36{feature} draw nav

+++ app/src/main/java/com/boetchain/bitcoinnode/worker/service/PeerManagementService.java
@@ -50,29 +50,49 @@ public class PeerManagementService extends Service {
      */
     private final IBinder binder = new LocalBinder();
 
+    private DnsSeedDiscoveryThread dnsSeedDiscoveryThread;
+
     private List<PeerCommunicatorThread> peerCommunicatorThreads = new ArrayList();
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
         Lawg.i("onStartCommand");
-        new PeerBroadcaster(this, new Peer(App.monitoringPeerIP)).broadcastLogAll("Bitcoin Service Starting...", ChatLog.TYPE_NEUTRAL);
-        isRunning = true;
 
-        Peer.deleteAll(Peer.class);
+        if (!isRunning) {
 
-        LocalBroadcastManager.getInstance(this).registerReceiver(localBroadcastReceiver, new IntentFilter(ACTION_DNS_SEED_DISCOVERY_COMPLETE));
-        LocalBroadcastManager.getInstance(this).registerReceiver(localBroadcastReceiver, new IntentFilter(PeerBroadcaster.ACTION_PEER_CONNECTED));
-        LocalBroadcastManager.getInstance(this).registerReceiver(localBroadcastReceiver, new IntentFilter(PeerBroadcaster.ACTION_PEER_DISCONNECTED));
+            isRunning = true;
 
-        peerPool = Peer.getPeerPool();
+            Lawg.i("Bitcoin Service Starting...");
+
+            Peer.deleteAll(Peer.class);
+
+            LocalBroadcastManager.getInstance(this).registerReceiver(localBroadcastReceiver, new IntentFilter(ACTION_DNS_SEED_DISCOVERY_COMPLETE));
+            LocalBroadcastManager.getInstance(this).registerReceiver(localBroadcastReceiver, new IntentFilter(PeerBroadcaster.ACTION_PEER_CONNECTED));
+            LocalBroadcastManager.getInstance(this).registerReceiver(localBroadcastReceiver, new IntentFilter(PeerBroadcaster.ACTION_PEER_DISCONNECTED));
+
+            findPeersAndConnect();
+        } else {
+
+            if (getConnectedPeers().size() <= 0) {
+                findPeersAndConnect();
+            }
+        }
+
+        return START_STICKY;
+    }
+
+    /**
+     * This will find peers to connect to either through the peer pool or by getting peers
+     * using the DNS discovery
+     */
+    private void findPeersAndConnect() {
+
         if (peerPool.size() == 0) {
             startDnsSeedDiscovery();
         } else {
             disconnectFromPeers();
             connectToPeers();
         }
-
-        return START_STICKY;
     }
 
     /**
@@ -90,10 +110,13 @@ public class PeerManagementService extends Service {
 
     /**
      * Starts the look up process to find initial peers or seeds to connect to.
+     * We only want one of these running at a time.
      */
     private void startDnsSeedDiscovery() {
-        Lawg.i("startDnsSeedDiscovery");
-        new DnsSeedDiscoveryThread(this).start();
+        if (dnsSeedDiscoveryThread != null && !dnsSeedDiscoveryThread.isRunning()) {
+            dnsSeedDiscoveryThread = new DnsSeedDiscoveryThread(this);
+            dnsSeedDiscoveryThread.start();
+        }
     }
 
     /**
@@ -167,13 +190,18 @@ public class PeerManagementService extends Service {
         disconnectFromPeers();
         LocalBroadcastManager.getInstance(this).unregisterReceiver(localBroadcastReceiver);
 
-        isRunning = false;
+        if (dnsSeedDiscoveryThread != null) {
+            dnsSeedDiscoveryThread.interrupt();
+        }
+
         new PeerBroadcaster(this, new Peer(App.monitoringPeerIP)).broadcastLogAll("Bitcoin Service Shutting down...", ChatLog.TYPE_NEUTRAL);
         Lawg.i("onDestroy");
 
         Peer.deleteAll(Peer.class);
 
         killPeerCommunicatorThreads();
+
+        isRunning = false;
     }
 
     /**
